mcp/requirements.txt
fastapi
uvicorn
akshare
pandas
pydantic
python-dotenv
mcp